my_differences_helper.php
This is a function which I found very practical to handle checkbox updates brought on the wrath of my project manager who called it "unreadable, not maintainable, not expandable and only good for elements with ids".

I thought it was a quick and handy way to identify the work needed and iterate through records that required the same operation applied. Was my approach wrong?